Fix textview with link implementations
