+++ corekit/src/main/res/values/strings.xml
@@ -143,6 +143,7 @@
     <string name="bca_bank_transfer">BCA</string>
     <string name="permata_bank_transfer">Permata</string>
     <string name="all_bank_transfer">Other Banks</string>
+    <string name="bni_bank_transfer">BNI</string>
 
 
     <string name="orderId">Order ID</string>
