+++ uikit/src/main/res/layout/fragment_instruction_bca_klik.xml
@@ -82,7 +82,7 @@
                 android:layout_height="wrap_content"
                 android:text="@string/number_instruction3" />
 
-            <com.midtrans.sdk.uikit.widgets.DefaultTextView
+            <com.midtrans.sdk.uikit.widgets.DefaultTextViewWithLink
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:background="@color/light_gray"
