+++ uikit/src/main/res/layout/fragment_instruction_klikbca.xml
@@ -58,7 +58,7 @@
                 android:layout_height="wrap_content"
                 android:text="@string/number_instruction2" />
 
-            <com.midtrans.sdk.uikit.widgets.DefaultTextView
+            <com.midtrans.sdk.uikit.widgets.DefaultTextViewWithLink
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:text="@string/instruction_klikbca2"
