+++ uikit/src/main/res/layout/fragment_klik_bca_payment.xml
@@ -73,10 +73,9 @@
                 android:layout_height="wrap_content"
                 android:text="@string/number_instruction2" />
 
-            <com.midtrans.sdk.uikit.widgets.DefaultTextView
+            <com.midtrans.sdk.uikit.widgets.DefaultTextViewWithLink
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-
                 android:text="@string/klikbca_instruction_2"
                 android:textColor="@color/black"
                 android:textSize="@dimen/secondary_text_size" />
